move createImposterToken to relay
