+++ src/server/graphql/rootMutation.js
@@ -1,5 +1,4 @@
 import {GraphQLObjectType} from 'graphql';
-import user from 'server/graphql/models/User/userMutation';
 import acceptTeamInviteEmail from 'server/graphql/mutations/acceptTeamInviteEmail';
 import acceptTeamInviteNotification from 'server/graphql/mutations/acceptTeamInviteNotification';
 import addAgendaItem from 'server/graphql/mutations/addAgendaItem';
@@ -60,18 +59,19 @@ import updateTeamName from 'server/graphql/mutations/updateTeamName';
 import createFirstTeam from 'server/graphql/mutations/createFirstTeam';
 import removeOrgUser from 'server/graphql/mutations/removeOrgUser';
 import createOrgPicturePutUrl from 'server/graphql/mutations/createOrgPicturePutUrl';
-
-const rootFields = Object.assign({},
-  user
-);
+import addFeatureFlag from 'server/graphql/mutations/addFeatureFlag';
+import createImposterToken from 'server/graphql/mutations/createImposterToken';
+import createUserPicturePutUrl from 'server/graphql/mutations/createUserPicturePutUrl';
+import updateUserWithAuthToken from 'server/graphql/mutations/updateUserWithAuthToken';
+import updateUserProfile from 'server/graphql/mutations/updateUserProfile';
 
 export default new GraphQLObjectType({
   name: 'Mutation',
   fields: () => ({
-    ...rootFields,
     acceptTeamInviteEmail,
     acceptTeamInviteNotification,
     addAgendaItem,
+    addFeatureFlag,
     addGitHubRepo,
     addOrg,
     addProvider,
@@ -83,10 +83,12 @@ export default new GraphQLObjectType({
     cancelTeamInvite,
     clearNotification,
     connectSocket,
+    createImposterToken,
     createFirstTeam,
     createGitHubIssue,
     createOrgPicturePutUrl,
     createProject,
+    createUserPicturePutUrl,
     deleteProject,
     disconnectSocket,
     editProject,
@@ -128,6 +130,8 @@ export default new GraphQLObjectType({
     updateCheckInQuestion,
     updateProject,
     updateTeamName,
+    updateUserProfile,
+    updateUserWithAuthToken,
     upgradeToPro
   })
 });
